Move ammo and item style to css file
With this commit the ammo and items objects are reduce to simply track amount/acquired state so there are no longer a need to store it in a value property.

To bridge the css selector, js identifier gap (dash vs underscore word separators) a pair of set/get properties are added to the ammo object.